+++ src/main/java/me/night0721/lilase/config/ConfigHandler.java
@@ -28,8 +28,9 @@ public class ConfigHandler {
         if (hasNoKey("OnlySniper")) setBoolean("OnlySniper", false);
         if (hasNoKey("GUI")) setBoolean("GUI", true);
         if (hasNoKey("GUI_COLOR")) setInt("GUI_COLOR", 0x003153);
-        if (hasNoKey("DEBUG")) setBoolean("DEBUG", false);
-        if (hasNoKey("RELIST_TIMEOUT")) setInt("RELIST_TIMEOUT", 1500);
+        if (hasNoKey("Debug")) setBoolean("Debug", false);
+        if (hasNoKey("RelistTimeout")) setInt("RelistTimeout", 1500);
+        if (hasNoKey("AuctionLength")) setInt("AuctionLength", 2);
     }
 
     public JsonObject readConfigFile() {