+++ src/main/java/me/night0721/lilase/features/claimer/ClaimerState.java
@@ -1,0 +1,5 @@
+package me.night0721.lilase.features.claimer;
+
+public enum ClaimerState {
+    NONE, WALKING_TO_FIRST_POINT, WALKING_INTO_AUCTION_HOUSE, OPENING, START
+}