+++ src/main/java/me/night0721/lilase/features/flipper/FlipperState.java
@@ -1,5 +1,5 @@
 package me.night0721.lilase.features.flipper;
 
 public enum FlipperState {
-    WALKING_TO_FIRST_POINT, WALKING_INTO_AUCTION_HOUSE, BUYING, START, NONE
+    WALKING_TO_FIRST_POINT, WALKING_INTO_AUCTION_HOUSE, BUYING, TIME, START, NONE
 }