+++ src/main/java/me/night0721/lilase/features/pageflipper/PageFlipperState.java
@@ -1,4 +1,4 @@
-package me.night0721.lilase.features.sniper;
+package me.night0721.lilase.features.pageflipper;
 
 public enum PageFlipperState {
     NONE, WALKING_TO_FIRST_POINT, WALKING_INTO_AUCTION_HOUSE, OPENING, START